requirements.txt
Django==2.2.28
django-cors-headers==3.2.0
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
esdk-obs-python==3.20.11
icalendar==4.0.9
mysqlclient==1.4.6
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.4
requests==2.31.0
wget==3.2
